# domin: set placement rule of partition table in batch