<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #43070 

Problem Summary:

### What is changed and how it works?

create a table with 1000 partitions

```sql
create table t (idx int, b int) 
PARTITION BY RANGE (idx) (
    PARTITION p0 VALUES LESS THAN (10),
    PARTITION p1 VALUES LESS THAN (20),
    PARTITION p2 VALUES LESS THAN (30),
    ......
    PARTITION p998 VALUES LESS THAN (9990),
    PARTITION p999 VALUES LESS THAN (10000)
);
```

before

```sql
mysql> truncate table t;
Query OK, 0 rows affected (1.22 sec)

mysql> alter table t set tiflash replica 1;
Query OK, 0 rows affected (2.06 sec)

mysql> truncate table t;
Query OK, 0 rows affected (5.67 sec)

mysql> truncate table t;
Query OK, 0 rows affected (7.69 sec)

mysql> alter table t set tiflash replica 0;
Query OK, 0 rows affected (5.80 sec)

mysql> truncate table t;
Query OK, 0 rows affected (1.26 sec)
```

 after

```sql
mysql> truncate table t;
Query OK, 0 rows affected (1.22 sec)

mysql> alter table t set tiflash replica 1;
Query OK, 0 rows affected (0.12 sec)

mysql> truncate table t;
Query OK, 0 rows affected (1.33 sec)

mysql> truncate table t;
Query OK, 0 rows affected (1.27 sec)

mysql> alter table t set tiflash replica 0;
Query OK, 0 rows affected (0.12 sec)

mysql> truncate table t;
Query OK, 0 rows affected (1.24 sec)
```


1. call `/config/rules/batch` instead of `/config/rule` for `ConfigureTiFlashPDForPartitions`
2. add a new API `/regions/accelerate-schedule/batch` in PD https://github.com/tikv/pd/pull/6327

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```